:truck: Rename error popup component to message popup
