+++ client/src/components/message-popup/MessagePopup.scss
@@ -1,13 +1,13 @@
 /**
  * @ Author: Jbristhuille
  * @ Create Time: 2022-12-24 15:00:00
- * @ Description: Error popup styles
+ * @ Description: Message popup styles
  */
 
 /* SUMMARY
 */
 
-.wcs-error-popup {
+.wcs-message-popup {
     color: white;
     background: var(--danger-color);
 }
